:recycle: Focus repo on just stm32f1
Attempting to support more than one target family in a single repo will
make versioning complicated for the consumers and for the library
managers, specifically, making breaking changes on a single target
family, but having to bump the version for the whole package. This can
confuse consumers and make it hard to reason about version changes.

:arrow_up: bump version to 0.3.0